OgmaJL API

OgmaJL.AW — Type

weight values AW

OgmaJL.FocalMatrix — Method

FocalMatrix(isize::XIO3, osize::XIO3, radius::Int)

Creates a FocalMatrix with given input sizes, output sizes and window radius.

Examples

import Random # hide
Random.seed!(1234) # hide
julia> f1 = FocalMatrix(XIO3(3, 4, 3), XIO3(4, 5, 3), 1)
FocalMatrix(XIO3[3×4×3], XIO3[4×5×3], 1, UInt16[1 2; 1 3; 2 3; 2 3], UInt16[1 2; 1 3; … ; 2 4; 3 4], UInt32[1 37 … 145 199; 235 289 … 451 532; 586 622 … 730 784; 820 856 … 964 1018], Float32[0.59084463, 0.76679707, 0.5662374, 0.46008533, 0.7940257, 0.8541466, 0.20058604, 0.29861426, 0.24683718, 0.5796722  …  0.587723, 0.42177233, 0.4241981, 0.3947768, 0.6542789, 0.11622245, 0.7428318, 0.48858702, 0.36625502, 0.7665335], 1054)

OgmaJL.SparseFocalEncoder — Method

SparseFocalEncoder(isizes::Vector{XIO3}, osize::XIO3, radii::Vector{Int})

Creates a SparseFocalEncoder with given input sizes, output size and window radii.

Examples

import Random # hide
Random.seed!(1234) # hide
julia> e = SparseFocalEncoder([XIO3(3, 4, 3), XIO3(5, 3, 2)], XIO3(4, 5, 3), [1, 1])

OgmaJL.XIO — Type

size index XIO

OgmaJL.XIO3 — Type

XIO3 stores x, y, z values of type XIO

OgmaJL.XIO3M — Type

mutable version of XIO3

OgmaJL.XW — Type

weight index XW

OgmaJL.forward! — Method

forward!(m::FocalMatrix, ins::Vector{AW}, outs::Vector{AW})

Runs a FocalMatrix forward with given inputs ins, fills outs with activations.

Examples

import Random # hide
Random.seed!(1234) # hide
m = FocalMatrix(XIO3(3, 4, 3), XIO3(4, 5, 3), 1);
Random.seed!(56789) # hide
ins = AW.(rand([-1,1], 3 * 4 * 3) .* rand(3 * 4 * 3))
outs = zeros(AW, 4 * 5 * 3)

julia> forward!(f, ins, outs)
60-element Array{Float32,1}:
  0.2664309
  1.8382773
 -0.067347586
  1.5660199
  0.26483107
  2.2840357
  0.71645874
  1.9513111
  2.453909
  0.20646897
  0.8714455
  2.0525088
  0.02235264
  ⋮
  2.767889
  1.1482695
  1.3052162
  1.0694035
  3.3695157
  1.2626712
  2.4824812
  0.7995742
  0.70120615
  0.70984924
  1.0287566
  0.99003124

OgmaJL.forward! — Method

forward!(e::SparseFocalEncoder, ins::Vector{Vector{AW}}, outs::Vector{AW})

Runs a SparseFocalEncoder forward with given inputs ins (copied into e.inputs), starting e.outputs .= outs, and fills e.outputs with activations.

Examples

import Random # hide
Random.seed!(34567)
osize = XIO3(4, 5, 3)
isizes = [XIO3(3, 4, 3), XIO3(5, 3, 2)]
e = SparseFocalEncoder(isizes, osize, [1, 1])
Random.seed!(89123)
outs = zero_array(AW, osize)
ins = [AW.(-1 .+ (2 .* rand(isize.x * isize.y * isize.z))) for isize in isizes]

julia> forward!(e, ins, outs)
60-element Array{Float32,1}:
  1.9380938
  1.412922
  1.4253737
  1.7408687
 -0.26759037
  0.8767195
  1.7407343
 -0.17168805
  0.4381518
  1.6412987
  1.9755627
  1.2878081
  2.071682
  ⋮
  0.26863363
  0.8956803
  0.7418641
  0.7577775
  1.1317881
  0.64095104
  0.995874
  1.9518139
  1.7899414
  1.6157343
  1.5994829
  1.4046599

OgmaJL.forward! — Method

forward!(e::SparseFocalEncoder)

Runs a SparseFocalEncoder e forward using its own inputs and mutating its outputs tensors.

OgmaJL.locateWeight — Method

locateWeight(m::FocalMatrix, index::XW)
locateWeight(m::FocalMatrix, index::Int) = locateWeight(m, XW(index))

Using m.osize dimensions, finds the input position [ix iy iz] and the output position [ox oy oz] for the given weight index.

Examples

m = FocalMatrix(XIO3(8,10,3), XIO3(5,4,3), 3);

julia> i, o = locateWeight(m, 4142)
(XIO3[4×5×2], XIO3[4×3×1])

OgmaJL.sparse — Method

sparse(m::FocalMatrix, outs::Vector{AW})

Using m.osize dimensions, chooses the top entry in the z-column for each x-y position in outs, returning a one-hot tensor (zero elsewhere) and an array of winner z-indices for each x-y position.

Examples

import Random # hide
Random.seed!(1234) # hide
m = FocalMatrix(XIO3(3, 4, 3), XIO3(4, 5, 3), 1);
Random.seed!(56789) # hide
ins = AW.(rand([-1,1], 3 * 4 * 3) .* rand(3 * 4 * 3))
outs = zeros(AW, 4 * 5 * 3)
outs = forward!(f, ins, outs)
julia> sparseouts, tops = sparse(f, outs)
(Float32[0.0, 0.0, 5.414768, 0.0, 0.0, 3.873893, 5.16832, 0.0, 0.0, 4.7478986  …  0.0, 7.7735734, 0.0, 0.0, 0.0, 0.0, 7.5303793, 4.5096736, 0.0, 0.0], UInt16[3, 3, 1, 1, 2, 3, 2, 2, 1, 2, 1, 3, 3, 1, 1, 2, 1, 1, 3, 1])

OgmaJL.sparse — Method

sparse(osize::XIO3, outs::Vector{AW})

Using osize dimensions, chooses the top entry in the z-column for each x-y position in outs, returning a one-hot tensor (zero elsewhere) and an array of winner z-indices for each x-y position.

Examples

import Random # hide
Random.seed!(1234) # hide
osize = XIO3(1234)
outs = rand_array(AW, osize)
julia> sparseouts, tops = sparse(osize, outs)
(Float32[0.0, 0.0, 0.5250573, 0.6120098, 0.0, 0.0, 0.0, 0.576082, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.982502, 0.4270338, 0.0, 0.0, 0.5586466, 0.0, 0.0], UInt16[3, 1, 2, 3, 3, 2, 2, 3, 1, 3, 1, 1, 3, 3, 1, 3, 1, 3, 1, 1])