The SparseFocalEncoder locally focussed encoder type

The SparseFocalEncoder connects a number of input tensors to a single hidden activation tensor, using an appropriate number of FocalMatrix weight matrices.

abstract type AbstractEncoder end

struct SparseFocalEncoder <: AbstractEncoder
    inputs::Vector{Vector{AW}}
    outputs::Vector{AW}
    weight_tensors::Vector{FocalMatrix}
end

There are three fields for the type.

inputs

A vector of input tensors.

outputs

The output activation tensor.

weight_tensors

A Vector of FocalMatrix, one for each input tensor.

Constructors

OgmaJL.SparseFocalEncoder — Type

SparseFocalEncoder(isizes::Vector{XIO3}, osize::XIO3, radii::Vector{Int})

Creates a SparseFocalEncoder with given input sizes, output size and window radii.

Examples

import Random # hide
Random.seed!(1234) # hide
julia> e = SparseFocalEncoder([XIO3(3, 4, 3), XIO3(5, 3, 2)], XIO3(4, 5, 3), [1, 1])

SparseFocalEncoder methods

There are two methods exported.

forward!

sparse

Missing docstring.

Missing docstring for forward!(::SparseFocalEncoder). Check Documenter's build log for details.

Missing docstring.

Missing docstring for forward!(::SparseFocalEncoder, ::Vector{Vector{AW}}, ::Vector{AW}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sparse(::XIO3, ::Array{AW}). Check Documenter's build log for details.