The FocalMatrix locally focussed weight matrix type

The FocalMatrix is a weight matrix which connects an input tensor ins of size isize (isize.x×isize.y×isize.z) to an output tensor outs of size osize (osize.x×osize.y×osize.z), where each output receives input from units in an x-y window within radius units of the input unit "below" the x-y of the output unit.

struct FocalMatrix
    isize::XIO3
    osize::XIO3
    radius::XIO
    xs::Array{XIO, 2}
    ys::Array{XIO, 2}
    idxs::Array{XW, 2}
    ws::Array{AW, 1}
    nWs::XW
end

There are nine fields for the type.

isize

The isize field specifies the dimensions of the layer's inputs (or visible layer) in x, y and z.

osize

The osize field specifies the dimensions of the layer's outputs (or hidden layer) in x, y and z.

radius

The radius field specifies the radius of a square x-y window in the inputs which are connected to any unit in the output layer.

xs (derived)

The xs field contains the minimum and maximum x edges of the input window for each x-y of the outputs. This field is computed by the external constructor below.

ys (derived)

The ys field contains the minimum and maximum y edges of the input window for each x-y of the outputs. This field is computed by the external constructor below.

idxs (derived)

The idxs field contains the starting weight index for a given output x-y-0 position. This field is computed by the external constructor below.

ws (derived)

The ws field contains the randomly initialised weight matrix. This field is computed by the external constructor below.

nWs (derived)

The nWs field is the size of the weight matrix. This field is computed by the external constructor below.

Constructors

OgmaJL.FocalMatrix — Type

FocalMatrix(isize::XIO3, osize::XIO3, radius::Int)

Creates a FocalMatrix with given input sizes, output sizes and window radius.

Examples

import Random # hide
Random.seed!(1234) # hide
julia> f1 = FocalMatrix(XIO3(3, 4, 3), XIO3(4, 5, 3), 1)
FocalMatrix(XIO3[3×4×3], XIO3[4×5×3], 1, UInt16[1 2; 1 3; 2 3; 2 3], UInt16[1 2; 1 3; … ; 2 4; 3 4], UInt32[1 37 … 145 199; 235 289 … 451 532; 586 622 … 730 784; 820 856 … 964 1018], Float32[0.59084463, 0.76679707, 0.5662374, 0.46008533, 0.7940257, 0.8541466, 0.20058604, 0.29861426, 0.24683718, 0.5796722  …  0.587723, 0.42177233, 0.4241981, 0.3947768, 0.6542789, 0.11622245, 0.7428318, 0.48858702, 0.36625502, 0.7665335], 1054)

FocalMatrix methods

There are three methods exported.

forward!

sparse

locateWeight