The Hierarchy sparse predictive hierarchy type

The Hierarchy contains a number of layers of encoders and decoders, arranged in a ladder.

struct Hierarchy
    inputSizes
    hiddenSizes
    radii
    tickLengths
    encoders
    decoders
end

There are six fields for the type.

inputSizes

A vector of XIO3 describing the shape of the inputs.

hiddenSizes

A vector of XIO3 describing the shapes of the hidden layers.

radii

A vector of vectors describing the feedforward window radii for each visible layer.

tickLengths

A vector of "clocks" for each layer.

encoders

The encoders for each layer.

decoders

The predictive decoders for each layer.

Constructors

OgmaJL.Hierarchy — Type

Hierarchy(inputSizes::Vector{XIO3}, hiddenSizes::Vector{XIO3}, radii::Vector{Vector{Int}}, tickLengths::Vector{Int})

Creates a Hierarchy with given input sizes, hidden layer sizes, window radii and tick lengths.

Examples

import Random # hide
Random.seed!(34567)

inSizes = [XIO3(3, 4, 3), XIO3(5, 3, 2)]
hidSizes = [XIO3(5, 5, 3), XIO3(4, 4, 2), XIO3(3, 3, 1)]
radii = [[1, 1], [2, 2], [1, 1]]
tickLengths = [1, 2, 2]

julia> h = Hierarchy(inSizes, hidSizes, radii, tickLengths)

Hierarchy methods

There are two methods exported.

forward!